request to connect to another connector(s)
I think it would be nice to be able to send a request (invitation) to connect to another connector(s) via the moneyd-gui interface.
For example: send to everyone or send to a specific connector.
After the connector confirms the request, everyone on their side chooses the role, enters (if necessary) additional data and presses to confirm, after which all the settings are automatically added to the configuration file (s). Well, or almost so.